requirements.txt
torch
torchvision
pyyaml
numpy
opencv-python-headless
tensorboard
tqdm
scipy
imgstore